requirements.txt
anthropic>=0.39.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
